olfs: Retired old BesNodeCache and NodeTransaction. Renamed new NodeCache to BesNodeCache Reduced size of locked block in (Bes)NodeTransaction. Dropped unused code from NodeTransaction.
