+++ src/opendap/bes/caching/LRUCache.java
@@ -16,7 +16,6 @@ class LRUCache<K,V> extends LinkedHashMap<K,V> {
         this.maxEntries = maxEntries;
     }
 
-
     @Override
     protected boolean removeEldestEntry(Map.Entry eldest) {
         return size() > maxEntries;
@@ -27,13 +26,12 @@ class LRUCache<K,V> extends LinkedHashMap<K,V> {
     private void show(){
         Logger log = LoggerFactory.getLogger(LRUCache.class);
         log.debug("Show...");
-        Iterator<Map.Entry<K,V>> it = entrySet().iterator();
-        while(it.hasNext()){
-            Map.Entry<K,V> entry = it.next();
-            log.debug("{}:{}",entry.getKey(),entry.getValue());
+        for (Map.Entry<K,V> entry : entrySet()) {
+            log.debug("{}:{}", entry.getKey(), entry.getValue());
         }
-
     }
+
+
 
     public static void main(String[] args){
         Logger log = LoggerFactory.getLogger(LRUCache.class);
@@ -49,31 +47,29 @@ class LRUCache<K,V> extends LinkedHashMap<K,V> {
         for(int i=0; i<10 ; i++){
             cache.put(Integer.toString(i), System.nanoTime()-start);
         }
-
         cache.show();
 
         log.debug("Retrieving values");
         for(int i=1; i<10 ; i+=2){
             cache.get(Integer.toString(i));
         }
-
+        cache.show();
 
         log.debug("Loading values");
         for(int i=10; i<30 ; i++){
             cache.put(Integer.toString(i), System.nanoTime()-start);
         }
-
         cache.show();
 
         log.debug("Retrieving values");
         for(int i=0; i<20 ; i+=2){
             cache.get(Integer.toString(i));
         }
+
         log.debug("Loading values");
         for(int i=0; i<5 ; i++){
             cache.put(Integer.toString(i), System.nanoTime()-start);
         }
-
         cache.show();
     }
 }
