+++ src/opendap/bes/caching/NodeTransaction.java
@@ -3,12 +3,9 @@ package opendap.bes.caching;
 import opendap.namespaces.BES;
 import org.jdom.Document;
 import org.jdom.Element;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.concurrent.atomic.AtomicLong;
 
 /**
  * This class is used to wrap whatever object is being cached along with data used to
@@ -17,19 +14,12 @@ import java.util.concurrent.atomic.AtomicLong;
  * This is not an autonomous operation and is tightly coupled with code in "BesNodeCache.getNode()" to
  * ensure that the ordering remains correct.
  */
-class NodeTransaction implements Comparable  {
-
-
-    private static final AtomicLong counter = new AtomicLong(0);
-
-    private Logger log;
+class NodeTransaction  {
 
     private Document request;
     private Object response;
-    private long lastUpdateTime;
-    private long lastAccessedTime;
+    private long timeCreated;
     private String key;
-    private long serialNumber;
 
 
     /**
@@ -43,10 +33,8 @@ class NodeTransaction implements Comparable  {
      *                 practice this is either an instance of Document or BESError.
      */
     NodeTransaction(String key, Document request, Object response){
-        log = LoggerFactory.getLogger(getClass());
         this.key = key;
         this.request = (Document)request.clone();
-
 
         // Dump the timeout context from the request.
         List list = this.request.getRootElement().getChildren("setContext", BES.BES_NS);
@@ -63,101 +51,28 @@ class NodeTransaction implements Comparable  {
         }
 
         this.response = response;
-        lastUpdateTime = System.nanoTime();
-        lastAccessedTime = lastUpdateTime;
-        serialNumber = counter.getAndIncrement();
+        timeCreated = System.nanoTime();
     }
 
     public String getKey(){
         return key;
     }
 
-    long getLastAccessedTime(){
-        return lastAccessedTime;
+
+    long getTimeCreated() {
+        return timeCreated;
     }
-
-    void updateAccessedTime() {
-        lastAccessedTime = System.nanoTime();
-    }
-
-
-
-    long getLastUpdateTime() {
-        return lastUpdateTime;
-    }
-
 
     public Object getResponse(){
         return response;
     }
+
     public void setResponse(Object response) {
         this.response = response;
     }
 
-
     public Document getRequest(){
         return (Document) request.clone();
-    }
-
-    /**
-     * The evaluation is based on the last accessed time (firstly) and the serial number of the
-     * CatalogTransaction (secondly). If the last accessed times of two objects are the same
-     * (unlikely but possible) then the serial numbers are used to determine the hierarchy/ranking/relation
-     * @param o object (CatalogTransaction) to be compared
-     * @return -1 if this object is older than Object o, 0 is they are the same, and 1 if this object
-     * is newer than Object o
-     */
-    @Override
-    public int compareTo(Object o) {
-        if (!(o instanceof NodeTransaction))
-            throw new ClassCastException("An instance of a NodeTransaction object was expected.");
-        NodeTransaction that = (NodeTransaction) o;
-        if(this==that)
-            return 0;
-
-        if(this.lastAccessedTime == that.lastAccessedTime){
-            log.warn("compareTo() - Required object serial numbers to differentiate " +
-                    "instances. this: {} that: {}",this.serialNumber, that.serialNumber);
-            return (int) (this.serialNumber - that.serialNumber);
-        }
-
-
-        // Why return like this? Because the return value is an integer and the computation produces a long
-        // incorrect conversion (over/under flow) could change sign of result.
-        return (this.lastAccessedTime - that.lastAccessedTime)>0?1:-1;
-
-
-    }
-
-    /**
-     *
-     * @param object The object to be compared.
-     * @return True if object is a reference to this instance, or if they are
-     * functionally the same. False otherwise.
-     */
-    @Override
-    public boolean equals(Object object) {
-        if (this == object) return true;
-        if (!(object instanceof NodeTransaction)) return false;
-
-        NodeTransaction that = (NodeTransaction)object;
-
-        return ( (this.lastAccessedTime == that.lastAccessedTime) &&
-                (this.request == that.request)  &&
-                (this.response == that.response)
-        );
-
-    }
-
-    /**
-     *
-     * @return A hash code representing this object.
-     */
-    @Override
-    public int hashCode() {
-        int result = 73;
-        result += lastAccessedTime + (request ==null?0: request.hashCode()) + (response ==null?0: response.hashCode());
-        return result;
     }
 
 }
